chore(clippy): silence pre-existing warnings (-W clippy::all)
`cargo clippy --all-targets -- -W clippy::all` was emitting 56 warnings.
Fixes them all so PR-level lint stays at zero:

- input.rs tests: `#[allow(disallowed_methods)]` on the `tests` module;
  ClickState's internal API is legitimately exercised here.
- cookie/simulator.rs: `map_or(false, ..)` → `is_some_and`.
- career/logic.rs: collapse nested `if`s; alias `fn(&mut CareerState)` as
  `StrategyFn` to drop type_complexity warning.
- career/mod.rs: drop `+ 0` in test indexing.
- cafe/characters & gacha tests: use struct-init shorthand instead of
  `Default::default()` + field reassignment.
- cafe/input_handler.rs: move `dismiss_popup` ahead of the test module.
- cafe/render/hub.rs: `id < A || id >= B` → `!(A..B).contains(&id)`.
- metropolis/logic.rs: drop redundant `as usize`; convert stale doc comment
  to a regular comment to remove `empty_line_after_doc`.
- metropolis/render.rs: `#[allow(too_many_arguments)]` on `tile_span_1` /
  `tile_spans_2` (8 args; refactoring to a struct here would obscure the
  hot rendering path).
- metropolis/terrain.rs: `#[allow(needless_range_loop)]` on two tests where
  both `x` and `y` are used as 2D indices.
- time.rs: `total >= 9 && total <= 11` → `(9..=11).contains(&total)`.

No behavior changes; all 862 tests still pass.

+++ src/games/metropolis/logic.rs
@@ -857,7 +857,7 @@ fn white_collar_capacity_within(
                 continue;
             }
             let (nx, ny) = (nx as usize, ny as usize);
-            if matches!(city.tile(nx as usize, ny as usize), Tile::Built(Building::Office))
+            if matches!(city.tile(nx, ny), Tile::Built(Building::Office))
                 && workshop_is_active_with(city, nx, ny, connected)
             {
                 total += OFFICE_CAPACITY_CENTS;
@@ -3164,9 +3164,9 @@ mod tests {
         assert_eq!((tx, ty), (cx, cy));
     }
 
-    /// Outpost 派遣のテストは AI 評価関数 (`placement_value`) のテスト経由で
-    /// 担保される (旧 dispatch_outpost / best_outpost_placement は AI 統合により廃止)。
-    /// 「saturation 時に Outpost が高評価になる」性質は `placement_value_*` テストで確認。
+    // Outpost 派遣のテストは AI 評価関数 (`placement_value`) のテスト経由で
+    // 担保される。`saturation 時に Outpost が高評価になる` 性質は
+    // `placement_value_*` テストで確認している。
 
     /// Rock セルは隣接 Outpost が無いと start_construction が false を返す。
     #[test]
